Refactor to *Operations
