+++ java-vtl-script/src/main/java/kohl/hadrien/vtl/script/operations/RenameOperation.java
@@ -1,4 +1,4 @@
-package kohl.hadrien.vtl.script;
+package kohl.hadrien.vtl.script.operations;
 
 /*-
  * #%L
