+++ java-vtl-script/src/main/java/kohl/hadrien/vtl/script/visitors/ClauseVisitor.java
@@ -22,7 +22,7 @@ package kohl.hadrien.vtl.script.visitors;
 
 import com.google.common.collect.ImmutableMap;
 import kohl.hadrien.*;
-import kohl.hadrien.vtl.script.RenameOperation;
+import kohl.hadrien.vtl.script.operations.RenameOperation;
 
 import java.util.List;
 import java.util.Optional;
